backend/requirements.txt
# =========================
# API
# =========================
fastapi
uvicorn[standard]
pydantic

# =========================
# HTTP CLIENT
# =========================
httpx
requests

# =========================
# ENV
# =========================
python-dotenv

# =========================
# AI
# =========================
openai
google-generativeai

# =========================
# SECURITY SCANNERS
# =========================
bandit
semgrep

# =========================
# GITHUB
# =========================
gitpython
pyjwt

# =========================
# REPORT
# =========================
reportlab

# =========================
# DATABASE
# =========================
sqlalchemy
psycopg2-binary

# =========================
# BACKGROUND TASKS
# =========================
celery
redis